backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
redis
hiredis
websockets
python-dotenv
psycopg2-binary
bcrypt==3.2.0
pyjwt
